# api/OpenApi/mcp-proxy.yaml
openapi: 3.0.3
info:
  title: Technijian MCPX Admin API
  version: "1.0.0"
servers:
  - url: /
paths:
...
  /config/public:
    get:
      summary: Public boot config (non-secret)
      responses:
        '200':
          description: Public config
          content:
            application/json:
              schema:
                type: object
                properties:
                  azureAd:
                    type: object
                    properties:
                      tenantId: { type: string }
                      clientId: { type: string }
                      redirectUri: { type: string }
                      scope: { type: string }
                  sse:
                    type: object
                    properties:
                      heartbeatSeconds: { type: integer }
                  cors:
                    type: object
                    properties:
                      allowedOrigins:
                        type: array
                        items: { type: string }
        '204': { description: No content }
...